Split button and widget
