+++ resources/assets/coffee/_classes/nav2.coffee
@@ -30,7 +30,11 @@ class @Nav2
     $(window).off 'throttled-resize.nav2-center-popup'
 
     for popup in document.querySelectorAll('.js-nav2--centered-popup')
-      if popup.dataset.clickMenuId != @currentMenu
+      container = popup.closest('.js-click-menu')
+
+      continue if !container?
+
+      if container.dataset.clickMenuId != @currentMenu
         popup.classList.add 'hidden'
         continue
