+++ resources/assets/lib/notification-button.tsx
@@ -1,0 +1,99 @@
+/**
+ *    Copyright (c) ppy Pty Ltd <contact@ppy.sh>.
+ *
+ *    This file is part of osu!web. osu!web is distributed with the hope of
+ *    attracting more community contributions to the core ecosystem of osu!.
+ *
+ *    osu!web is free software: you can redistribute it and/or modify
+ *    it under the terms of the Affero GNU General Public License version 3
+ *    as published by the Free Software Foundation.
+ *
+ *    osu!web is distributed WITHOUT ANY WARRANTY; without even the implied
+ *    warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ *    See the GNU Affero General Public License for more details.
+ *
+ *    You should have received a copy of the GNU Affero General Public License
+ *    along with osu!web.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import * as _ from 'lodash';
+import { observer } from 'mobx-react';
+import NotificationWidget from 'notification-widget/main';
+import { NotificationContext } from 'notifications-context';
+import Worker from 'notifications/worker';
+import * as React from 'react';
+
+interface Props {
+  type?: string;
+  worker: Worker;
+}
+
+@observer
+export default class NotificationButton extends React.Component<Props> {
+  private menuId = `nav-notification-popup-${osu.uuid()}`;
+
+  render() {
+    if (!this.props.worker.isActive()) {
+      return null;
+    }
+
+    return (
+      <>
+        <button
+          className={this.buttonClass()}
+          data-click-menu-target={this.menuId}
+        >
+          <span className={this.mainClass()}>
+            <i className='fas fa-inbox' />
+            <span className='notification-icon__count'>
+              {this.unreadCount()}
+            </span>
+          </span>
+        </button>
+        <div
+          className='nav-click-popup js-click-menu'
+          data-click-menu-id={this.menuId}
+          data-visibility='hidden'
+        >
+          <NotificationContext.Provider value={{ isWidget: true }}>
+            <NotificationWidget extraClasses='js-nav2--centered-popup'/>
+          </NotificationContext.Provider>
+        </div>
+      </>
+    );
+  }
+
+  private buttonClass() {
+    let ret = 'js-click-menu nav-button';
+
+    if (this.props.type === 'mobile') {
+      ret += ' nav-button--mobile';
+    } else {
+      ret += ' nav-button--stadium';
+    }
+
+    return ret;
+  }
+
+  private mainClass() {
+    let ret = 'notification-icon';
+
+    if (this.props.worker.unreadCount > 0) {
+      ret += ' notification-icon--glow';
+    }
+
+    if (this.props.type === 'mobile') {
+      ret += ' notification-icon--mobile';
+    }
+
+    return ret;
+  }
+
+  private unreadCount() {
+    if (this.props.worker.hasData) {
+      return osu.formatNumber(this.props.worker.unreadCount);
+    } else {
+      return '...';
+    }
+  }
+}
